Refactor command configuration and info classes: Remove old command_config.hpp and command_info.hpp files, and introduce new implementations in the commands directory for improved organization and clarity in command handling.
