+++ worm_picker_core/include/worm_picker_core/core/commands/command_info.hpp
